--- src/components/Card/card-body/card-body.css.ts
@@ -1,8 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const body = style({
-  padding: '1rem',
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '0.5rem',
-});
