--- src/components/Card/card-body/index.tsx
@@ -1,1 +1,0 @@
-export * from './card-body';
