--- src/components/Card/card-footer/card-footer.css.ts
@@ -1,7 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const footer = style({
-  padding: '1rem',
-  display: 'flex',
-  alignItems: 'center',
-});
