--- src/components/Card/card-footer/index.tsx
@@ -1,1 +1,0 @@
-export * from './card-footer';
