--- src/components/Card/card-header/card-header.css.ts
@@ -1,7 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const header = style({
-  padding: '1rem',
-  display: 'flex',
-  flexDirection: 'row',
-});
