--- src/components/Card/card-header/index.ts
@@ -1,1 +1,0 @@
-export * from './card-header';
