--- src/components/Card/card-image/card-image.css.ts
@@ -1,10 +1,0 @@
-import { createVar, style } from '@vanilla-extract/css';
-
-export const maxWidth = createVar();
-
-export const image = style({
-  width: '100%',
-  verticalAlign: 'top',
-  objectFit: 'cover',
-  maxWidth: maxWidth,
-});
