--- src/components/Card/card-image/index.ts
@@ -1,1 +1,0 @@
-export * from './card-image';
