--- src/components/Card/card-title/card-title.css.ts
@@ -1,6 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const title = style({
-  fontSize: '1.4rem',
-  fontWeight: 'bold',
-});
