--- src/components/Card/card-title/index.tsx
@@ -1,1 +1,0 @@
-export * from './card-title';
