--- src/components/Card/card/index.ts
@@ -1,1 +1,0 @@
-export * from './card';
