+++ src/components/Card/index.tsx
@@ -1,6 +1,1 @@
 export * from './card';
-export * from './card-header';
-export * from './card-title';
-export * from './card-body';
-export * from './card-footer';
-export * from './card-image';
